@@ -1,18 +1,35 @@
 using System.IO;
 using System.Net.Http;
+using System.Text;
 using System.Threading.Tasks;
+using System.Xml;
 using System.Xml.Serialization;
 using Newtonsoft.Json;
 
 namespace OsmIntegrator.Tools
 {
+  public class Utf8StringWriter : StringWriter
+  {
+    public override Encoding Encoding => Encoding.UTF8;
+  }
+
   public class SerializationHelper
   {
     public static string XmlSerialize<T>(T changeNode)
     {
+      XmlWriterSettings xmlSettings = new()
+      {
+        Indent = false,
+        NewLineHandling = NewLineHandling.None,
+        Encoding = Encoding.UTF8,
+        OmitXmlDeclaration = true,
+        NamespaceHandling = NamespaceHandling.OmitDuplicates
+      };
+
       XmlSerializer serializer = new XmlSerializer(typeof(T));
-      using StringWriter textWriter = new StringWriter();
-      serializer.Serialize(textWriter, changeNode);
+      using Utf8StringWriter textWriter = new();
+      using XmlWriter xmlWriter = XmlWriter.Create(textWriter, xmlSettings);
+      serializer.Serialize(xmlWriter, changeNode);
       return textWriter.ToString();
     }
 
