@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using OsmIntegrator.Database.Models;
+using OsmIntegrator.ApiModels.Tiles;
 using OsmIntegrator.Database.Models.Enums;
 
 namespace OsmIntegrator.ApiModels.Stops
@@ -19,11 +19,9 @@ namespace OsmIntegrator.ApiModels.Stops
     public StopType StopType { get; set; }
     public ProviderType ProviderType { get; set; }
     public Guid TileId { get; set; }
-    public Tile Tile { get; set; }
-    public bool OutsideSelectedTile { get; set; } = false;
+    public bool OutsideSelectedTile { get; set; }
     public int Version { get; set; }
     public string Changeset { get; set; }
-
-    public bool IsDeleted { get; set; } = false;
+    public bool IsDeleted { get; set; }
   }
 }
