@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Linq;
 using AutoMapper;
-using MimeKit;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -18,7 +17,6 @@ using Microsoft.AspNetCore.Cors;
 using OsmIntegrator.Database.Models;
 using OsmIntegrator.Database;
 using OsmIntegrator.ApiModels;
-using OsmIntegrator.Interfaces;
 using OsmIntegrator.Database.Models.Enums;
 using OsmIntegrator.ApiModels.Conversation;
 
@@ -48,8 +46,6 @@ namespace OsmIntegrator.Controllers
     private readonly IMapper _mapper;
     private readonly IStringLocalizer<ConversationController> _localizer;
 
-    private readonly IEmailService _emailService;
-
     private readonly IConfiguration _configuration;
 
     public ConversationController(
@@ -59,7 +55,6 @@ namespace OsmIntegrator.Controllers
         RoleManager<ApplicationRole> roleManager,
         ApplicationDbContext dbContext,
         IStringLocalizer<ConversationController> localizer,
-        IEmailService emailService,
         IConfiguration configuration
     )
     {
@@ -69,7 +64,6 @@ namespace OsmIntegrator.Controllers
       _roleManager = roleManager;
       _dbContext = dbContext;
       _localizer = localizer;
-      _emailService = emailService;
       _configuration = configuration;
     }
 
@@ -172,130 +166,17 @@ namespace OsmIntegrator.Controllers
       return Ok(response);
     }
 
-
-    [HttpPut("Approve/{conversationId}")]
-    [Authorize(Roles =
-        UserRoles.SUPERVISOR + "," +
-        UserRoles.COORDINATOR + "," +
-        UserRoles.ADMIN)]
-    public async Task<ActionResult> Approve(string conversationId)
-    {
-      DbConversation dbConversation = await _dbContext.Conversations
-        .Include(x => x.Messages)
-          .ThenInclude(y => y.User)
-        .FirstOrDefaultAsync(x => x.Id == Guid.Parse(conversationId));
-
-      ApplicationUser user = await _userManager.GetUserAsync(User);
-
-      if (dbConversation == null)
-      {
-        throw new BadHttpRequestException(_localizer["Selected conversation doesn't exist"]);
-      }
-
-      Message approvalMessage = new Message()
-      {
-        Id = new Guid(),
-        UserId = user.Id,
-        ConversationId = dbConversation.Id,
-        Status = NoteStatus.Approved,
-        CreatedAt = DateTime.Now
-      };
-      DbMessage dbMessage = _mapper.Map<DbMessage>(approvalMessage);
-      dbMessage.User = user;
-
-      dbConversation.Messages.Add(dbMessage);
-
-      _dbContext.SaveChanges();
-
-      return Ok(_localizer["Conversation approved successfully"]);
-    }
-
-    private void SendApprovedTileEmail(DbConversation conversation)
-    {
-      List<ApplicationUser> users = conversation.Messages
-        .FindAll(x => x.Status == NoteStatus.Created)
-        .Select(x => x.User)
-        .ToList();
-
-      foreach (ApplicationUser user in users)
-      {
-        Task task = Task.Run(() => _emailService.SendEmailAsync(ConversationResolvedMessageBuilder(user)));
-      }
-    }
-
-    private MimeMessage ConversationResolvedMessageBuilder(ApplicationUser user)
-    {
-      MimeMessage message = new MimeMessage();
-      message.From.Add(MailboxAddress.Parse(_configuration["Email:SmtpUser"]));
-      message.To.Add(MailboxAddress.Parse(user.Email));
-      message.Subject = _emailService.BuildSubject(_localizer["Report resolved"]);
-
-      BodyBuilder builder = new BodyBuilder();
-
-      builder.TextBody = $@"{_localizer["Hello"]} {user.UserName},
-{_localizer["Report that you submitted was resolved"]}
-{_emailService.BuildServerName(false)}
-{_localizer["Regards"]},
-{_localizer["OsmIntegrator Team"]},
-rozwiazaniadlaniewidomych.org
-      ";
-      builder.HtmlBody = $@"<h3>{_localizer["Hello"]} {user.UserName},</h3>
-<p>{_localizer["Report that you submitted was resolved"]}</p><br/>
-{_emailService.BuildServerName(true)}
-<p>{_localizer["Regards"]},</p>
-<p>{_localizer["OsmIntegrator Team"]},</p>
-<a href=""rozwiazaniadlaniewidomych.org"">rozwiazaniadlaniewidomych.org</a>
-      ";
-
-      message.Body = builder.ToMessageBody();
-
-      return message;
-    }
-
-    [HttpPut("Reject/{conversationId}")]
-    [Authorize(Roles =
-        UserRoles.SUPERVISOR + "," +
-        UserRoles.COORDINATOR + "," +
-        UserRoles.ADMIN)]
-    public async Task<ActionResult> Reject(string conversationId)
-    {
-      DbConversation dbConversation = await _dbContext.Conversations
-        .Include(x => x.Messages)
-        .FirstOrDefaultAsync(x => x.Id == Guid.Parse(conversationId));
-
-      ApplicationUser user = await _userManager.GetUserAsync(User);
-
-      if (dbConversation == null)
-      {
-        throw new BadHttpRequestException(_localizer["Selected conversation doesn't exist"]);
-      }
-
-      Message rejectMessage = new Message()
-      {
-        Id = new Guid(),
-        UserId = user.Id,
-        ConversationId = dbConversation.Id,
-        Status = NoteStatus.Rejected,
-        CreatedAt = DateTime.Now
-      };
-      DbMessage dbMessage = _mapper.Map<DbMessage>(rejectMessage);
-      dbMessage.User = user;
-
-      dbConversation.Messages.Add(dbMessage);
-
-      _dbContext.SaveChanges();
-      _dbContext.SaveChanges();
-
-      return Ok(_localizer["Conversation rejected successfully"]);
-    }
-
-
     /// <summary>
     /// Approves conversation.
     /// </summary>
     /// <param name="message">Message object.</param>
     /// <returns>Operation satuts.</returns>
     [HttpPut("Approve")]
+    [Authorize(Roles =
+        UserRoles.EDITOR + "," +
+        UserRoles.SUPERVISOR + "," +
+        UserRoles.COORDINATOR + "," +
+        UserRoles.ADMIN)]
     public async Task<ActionResult> Approve([FromBody] MessageInput messageInput)
     {
       DbConversation dbConversation = await _dbContext.Conversations
@@ -326,48 +207,6 @@ rozwiazaniadlaniewidomych.org
       _dbContext.SaveChanges();
 
       return Ok(_localizer["Conversation approved successfully"]);
-    }
-
-    /// <summary>
-    /// Rejects conversation.
-    /// </summary>
-    /// <param name="message">Message object.</param>
-    /// <returns>Operation satuts.</returns>
-    [HttpPut("Reject")]
-    [Authorize(Roles =
-        UserRoles.SUPERVISOR + "," +
-        UserRoles.COORDINATOR + "," +
-        UserRoles.ADMIN)]
-    public async Task<ActionResult> Reject([FromBody] MessageInput messageInput)
-    {
-      DbConversation dbConversation = await _dbContext.Conversations
-        .Include(x => x.Messages)
-        .FirstOrDefaultAsync(x => x.Id == messageInput.ConversationId);
-
-      ApplicationUser user = await _userManager.GetUserAsync(User);
-
-      if (dbConversation == null)
-      {
-        throw new BadHttpRequestException(_localizer["Selected conversation doesn't exist"]);
-      }
-
-      Message message = new Message()
-      {
-        Id = new Guid(),
-        UserId = user.Id,
-        ConversationId = dbConversation.Id,
-        Status = NoteStatus.Rejected,
-        CreatedAt = DateTime.Now,
-        Text = messageInput.Text,
-      };
-      DbMessage dbMessage = _mapper.Map<DbMessage>(message);
-      dbMessage.User = user;
-
-      dbConversation.Messages.Add(dbMessage);
-
-      _dbContext.SaveChanges();
-
-      return Ok(_localizer["Conversation rejected successfully"]);
     }
   }
 }
