@@ -24,22 +24,18 @@ namespace OsmIntegrator.Database.Models
     [Required]
     public DbStop GtfsStop { get; set; }
 
-    [Required]
-    public bool Imported { get; set; }
-
     public Guid? UserId { get; set; }
 
     public ApplicationUser User { get; set; }
 
     public ConnectionOperationType OperationType { get; set; }
 
-    public DateTime? UpdatedAt { get; set; }
-
     [Required]
     public DateTime CreatedAt { get; set; }
 
-    public Guid? ApprovedById { get; set; }
+    [Required]
+    public bool Exported { get; set; }
 
-    public ApplicationUser ApprovedBy { get; set; }
+    public DateTime? ExportedAt { get; set; }
   }
 }
