@@ -5,13 +5,6 @@ namespace OsmIntegrator.Interfaces
 {
     public interface IEmailService
     {
-        void Send(string to, string subject, string message);
-        void Send(MimeMessage message);
-        Task SendEmailAsync(string to, string subject, string message);
-        Task SendEmailAsync(MimeMessage message);
-
-        string BuildSubject(string subject);
-
-        string BuildServerName(bool useHtml);
+        Task SendEmailAsync(string to, string subject, MimeEntity message);
     }
 }
