@@ -8,8 +8,8 @@ namespace OsmIntegrator.Interfaces
   public interface IReportsFactory
   {
     void UpdateName(ReportStop reportStop, string name);
-    ReportTile Create(DbTile tile);
-    ReportStop CreateStop(ReportTile report, Node node, DbStop stop, ChangeAction action, bool reverted = false);
+    TileImportReport Create(DbTile tile);
+    ReportStop CreateStop(TileImportReport report, Node node, DbStop stop, ChangeAction action, bool reverted = false);
     void AddField(ReportStop reportStop, string name, string actualValue, string previousValue, ChangeAction action);
   }
 }
